refactor: Readability of UI code.
